api/radiam/requirements.txt
coverage
cryptography
django
django-cors-headers
django-oauth-toolkit
djangorestframework==3.9.0
djangorestframework_simplejwt
django-filter
django-mptt
django-nose
django-rest-passwordreset==1.0.0a3
dry-rest-permissions
sendgrid-django
pillow
psycopg2-binary
python-memcached
idna==2.7
elasticsearch-dsl>=6.0.0,<7.0.0
coreapi
pygments
markdown
gunicorn
